* upstream/main: (98 commits)
  [pre-commit.ci] pre-commit autoupdate (#6335)
  SPEC 0: drop py310 and support py313 (#6195)
  Better benchmarking Python version handling (#6329)
  Move loading and combine code into their own submodules. (#6321)
  Bump scitools/workflows from 2025.02.1 to 2025.02.2 (#6327)
  replaced reference from build to python build (#6324)
  [pre-commit.ci] pre-commit autoupdate (#6315)
  Cache Dask arrays created from `NetCDFDataProxy`s to speed up loading files with multiple variables (#6252)
  Bump scitools/workflows from 2025.02.0 to 2025.02.1 (#6313)
  [pre-commit.ci] pre-commit autoupdate (#6310)
  Bump scitools/workflows from 2025.01.5 to 2025.02.0 (#6306)
  Updated environment lockfiles (#6301)
  Improve speed of loading small NetCDF files (#6229)
  [pre-commit.ci] pre-commit autoupdate (#6298)
  Use cube chunks for weights in aggregations with smart weights (#6288)
  Updated environment lockfiles (#6296)
  Bump scitools/workflows from 2025.01.4 to 2025.01.5 (#6300)
  Bump scitools/workflows from 2025.01.3 to 2025.01.4 (#6295)
  Lazy rectilinear interpolator  (#6084)
  Revert "Fix broken link. (#6246)" (#6297)
  ...

+++ lib/iris/common/resolve.py
@@ -2358,16 +2358,16 @@ class Resolve:
             >>> resolver.map_rhs_to_lhs
             True
             >>> cube1.data.sum()
-            124652160.0
+            np.float32(124652160.0)
             >>> zeros.shape
             (240, 37, 49)
             >>> zeros.sum()
-            0.0
+            np.float32(0.0)
             >>> result = resolver.cube(zeros, in_place=True)
             >>> result is cube1
             True
             >>> cube1.data.sum()
-            0.0
+            np.float32(0.0)
 
         """
         from iris.cube import Cube
@@ -2592,9 +2592,9 @@ class Resolve:
                     source                      'Data from Met Office Unified Model 6.05'
             >>> Resolve().shape is None
             True
-            >>> Resolve(cube1, cube2).shape
+            >>> Resolve(cube1, cube2).shape  # doctest: +SKIP
             (240, 37, 49)
-            >>> Resolve(cube2, cube1).shape
+            >>> Resolve(cube2, cube1).shape  # doctest: +SKIP
             (240, 37, 49)
 
         """  # noqa: D214, D406, D407, D410, D411
